add notifications

+++ app/Http/Controllers/Api/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Mail\WebMail;
 use App\Jobs\SendEmailJob;
+use App\Models\Admin;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -17,9 +18,9 @@ class UserController extends Controller
      */
     public function index()
     {
-        $useDB = User::all();
-        $users = $useDB->map(fn($user) => $user->toArray());
-        return response()->json($users, 200);
+        $useDB = Admin::all();
+        $admins = $useDB->map(fn($user) => $user->toArray());
+        return response()->json($admins, 200);
     }
 
     /**
@@ -44,26 +45,32 @@ class UserController extends Controller
             'roles.*' => 'exists:roles,id',
         ]);
         $user = User::create([
-            'nom' => $validated['nom'],
-            'prenom' => $validated['prenom'],
             'email' => $validated['email'],
             'password' => bcrypt('123456'),
-            'phone' => $validated['phone']
         ]);
 
         if ($user) {
-            $user->roles()->attach($validated['roles']);
-            $data = [
-                'nom' => $user->nom,
-                'prenom' => $user->prenom,
-                'phone' => $user->phone,
-                "email" => $user->email,
-                "password" => "123456",
-                'sujet' => "Confirmatiom du compte"
-            ];
-            //  SendEmailJob::dispatch($user->email, $data);
-            Mail::to($user->email)->send(new WebMail($data));
-            return response()->json($user, 200);
+            $admin = Admin::create([
+                'nom' => $validated['nom'],
+                'prenom' => $validated['prenom'],
+                'phone' => $validated['phone'],
+                'user_id' => $user->id,
+            ]);
+            if ($admin) {
+                $admin->roles()->attach($validated['roles']);
+                $data = [
+                    'nom' => $admin->nom,
+                    'prenom' => $admin->prenom,
+                    'phone' => $admin->phone,
+                    "email" => $admin->users->email,
+                    "password" => "123456",
+                    'sujet' => "Confirmatiom du compte"
+                ];
+                //  SendEmailJob::dispatch($user->email, $data);
+                Mail::to($user->email)->send(new WebMail($data));
+                return response()->json($admin, 200);
+            }
+            return response()->json(false, 500);
         } else {
             return response()->json(false, 500);
         }
@@ -75,7 +82,7 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
-        $user = User::with('roles')->findOrFail($id);
+        $user = Admin::with('roles')->findOrFail($id);
         $roles = [];
         foreach ($user->roles as $role) {
             $roles[$role->id] = $role->nom;
@@ -112,15 +119,15 @@ class UserController extends Controller
             'roles' => 'required|array', // Assurez-vous que 'roles' est un tableau
             'roles.*' => 'exists:roles,id',
         ]);
-        $user = User::findOrFail($id);
+        $admin = Admin::findOrFail($id);
 
-        if ($user) {
-            $user->phone = $validated['phone'];
-            $user->nom = $validated['nom'];
-            $user->prenom = $validated['prenom'];
-            $user->roles()->detach();
-            $user->roles()->attach($validated['roles']);
-            $user->save();
+        if ($admin) {
+            $admin->phone = $validated['phone'];
+            $admin->nom = $validated['nom'];
+            $admin->prenom = $validated['prenom'];
+            $admin->roles()->detach();
+            $admin->roles()->attach($validated['roles']);
+            $admin->save();
             return response()->json(true, 200);
         }
         return response()->json($validated, 200);
@@ -131,10 +138,12 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        $user = User::findOrFail($id);
-        if ($user) {
-            $user->delete();
-
+        $admin = Admin::findOrFail($id);
+        if ($admin) {
+            $user = User::find($admin->users->id);
+            if ($user)
+                $user->delete();
+            $admin->delete();
             return response()->json(true, 200);
         }
 
@@ -143,31 +152,30 @@ class UserController extends Controller
 
     public function change_status(string $id)
     {
-        $user = User::find($id);
+        $admin = Admin::find($id);
 
-        if ($user->status == true) {
-            $user->status = false;
-        } else {
-            $user->status = true;
+        if ($admin) {
+            $user = User::find($admin->users->id);
+            $user->status = $user->status ? false : true;
+            $admin->status = $admin->status ? false : true;
+            $user->save();
+            $admin->save();
         }
 
-        $user->save();
-        return response()->json($user);
+
+        return response()->json(true, 200);
     }
 
     public function gerants()
     {
-        $gerants = User::whereHas('roles', function ($query) {
+        $gerants = Admin::where('status', true)->whereHas('roles', function ($query) {
             $query->where('nom', 'gérant');
         })->get();
 
-        $nouveauGerants = [];
-        foreach ($gerants as $gerant) {
-            $nouveauGerants[] = [
-                'id' => $gerant->id,
-                'nom' => "$gerant->prenom $gerant->nom"
-            ];
-        }
+        $nouveauGerants = $gerants->map(fn($gerant) => [
+            'id' => $gerant->id,
+            'nom' => "$gerant->prenom $gerant->nom"
+        ]);
         return response()->json($nouveauGerants, 200);
     }
 
@@ -179,17 +187,18 @@ class UserController extends Controller
         ]);
         $password = $request->password;
         $user = User::where("email", $request->email)->first();
-
-        if ($user && Hash::check($password, $user->password)) {
-            $token = $user->createToken("client-$user->email")->plainTextToken;
-            $userLogin = $user->toArray();
-            $userLogin["token"]=$token;
-            return response()->json($userLogin, 201);
+        if ($user && Hash::check($password, $user->password) && $user->status == true) {
+            $token = $user->createToken("admin-$user->email")->plainTextToken;
+            $admin = Admin::where('user_id', $user->id)->first();
+            $admin_login =  $admin->toArray();
+            $admin_login["token"] = $token;
+            return response()->json($admin_login, 201);
         }
         return response()->json(false, 500);
     }
 
-    public function logout (Request $request){
-        
+    public function logout(Request $request)
+    {
+
     }
 }
